% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_ale.R
\name{bayes_ale}
\alias{bayes_ale}
\title{Accumulated Local Effects for Bayesian Posterior Distributions}
\usage{
bayes_ale(
  x,
  f_hat,
  vars,
  k = 40,
  limits = c(0.025, 0.975),
  num_burn = 0,
  num_thin = 1,
  include_main_effects = TRUE,
  center = TRUE,
  f = NULL
)
}
\arguments{
\item{x}{Data frame of covariates used in the model-fitting process.}

\item{f_hat}{Function that returns a posterior distribution of predictions given covariates \code{x}. The
function should return a matrix with \code{nrow(x)} rows and \code{num_save} columns.}

\item{vars}{Character vector of covariate names to vary in partial dependence calculations.}

\item{k}{Numeric vector of unique values to evaluate each covariate in \code{vars} at. By default this value will be decreased to the number of unique values if necessary.}

\item{limits}{Numeric vector of length 2 indicating the quantiles used for the lower and upper bounds of the credible interval.}

\item{num_burn}{Number of posterior samples to skip when making predictions.}

\item{num_thin}{Thinning interval for posterior samples following \code{num_burn} samples.}

\item{include_main_effects}{Logical value indicating whether main effect ALEs should be added to
second-order ALE calculations. Has no effect when \code{length(vars)} is 1.}

\item{center}{Either a logical value indicating whether or not to vertically center the ALE
with respect to the average prediction, or 'median' to indicate centering should be done with
respect to the median x value(s).}

\item{f}{(Optional) function that returns the true prediction for a set of covariates. Useful when conducting simulation studies.}
}
\value{
A data frame of posterior mean estimates and credible intervals for
a predictor grid. For second-order ALE, plotting intervals are also returned
for custom plotting similar to \code{image}.
}
\description{
Compute first and second-order accumulated local effects (ALE) with credible
intervals.
}
