% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_pd.R
\name{bayes_pd}
\alias{bayes_pd}
\title{Partial Dependence for Bayesian Posterior Distributions}
\usage{
bayes_pd(
  x,
  f_hat,
  vars,
  k,
  k_vals = NULL,
  limits = c(0.025, 0.975),
  num_burn = 0,
  num_thin = 1,
  f = NULL
)
}
\arguments{
\item{x}{Data frame of covariates used in the model-fitting process.}

\item{f_hat}{Function that returns a posterior distribution of predictions given covariates \code{x}. The
function should return a matrix with \code{nrow(x)} rows and \code{num_save} columns.}

\item{vars}{Character vector of covariate names to vary in partial dependence calculations.}

\item{k}{Numeric vector of unique values to evaluate each covariate in \code{vars} at. By default this value will be decreased to the number of unique values if necessary.}

\item{k_vals}{Alternatively, a list of \code{length(vars)} numeric vectors, each representing the unique values to evaluate the corresponding variable at.}

\item{limits}{Numeric vector of length 2 indicating the quantiles used for the lower and upper bounds of the credible interval.}

\item{num_burn}{Number of posterior samples to skip when making predictions.}

\item{num_thin}{Thinning interval for posterior samples following \code{num_burn} samples.}

\item{f}{(Optional) function that returns the true prediction for a set of covariates. Useful when conducting simulation studies.}
}
\value{
A data frame of posterior mean estimates and 95% credible intervals.
}
\description{
Compute partial dependence statistics with credible intervals.
}
