% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdpd-package.R
\docType{package}
\name{pdpd-package}
\alias{pdpd}
\alias{pdpd-package}
\title{pdpd: Partial Dependence for Posterior Distributions}
\description{
Compute statistics to aid the interpretability of black-box Bayesian machine learning models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jacobenglert/pdpd}
  \item Report bugs at \url{https://github.com/jacobenglert/pdpd/issues}
}

}
\author{
\strong{Maintainer}: Jacob Englert \email{jacob.englert@emory.edu}

}
\keyword{internal}
